First in series of posts reworking posts and DSOs.
THIS BRANCH IS NOT GUARANTEED TO GREEN-BAR UNTIL MERGED BACK.
IT IS A WORK IN PROGRESS. YHBW.

Now that that's out of the way… in reorganising the DSO logic
decoupling persistence and publishing (see the discussion on Issue
#100), the newly-renamed BlogPostAdder seemed a sensible place to
start. Note that its `#execute` method no longer simply calls
`#publish` on the passed-in `post`; it calls a private
`#post_method_message` method that returns `:publish` for a `public`
post, and `:add_to_blog` for a `draft` post.

RSpec: 723 examples,12 failures (but see Issue #87)
RuboCop: 151 files inspected, no offenses detected